chore: extract actions panel helper types
