+++ packages/web/tests/helpers/actionsPanel.types.ts
@@ -1,0 +1,25 @@
+import type { PopulationConfig } from '@kingdom-builder/protocol';
+import type { createActionsPanelGame } from './actionsPanel';
+
+export interface ActionsPanelGameOptions {
+	populationRoles?: Array<Partial<PopulationConfig>>;
+	showBuilding?: boolean;
+	actionCategories?: {
+		population?: string;
+		basic?: string;
+		building?: string;
+	};
+	requirementBuilder?: (context: {
+		capacityStat: string;
+		populationPlaceholder: string;
+	}) => unknown[];
+	resourceKeys?: {
+		actionCost?: string;
+		upkeep?: string;
+	};
+	statKeys?: {
+		capacity?: string;
+	};
+}
+
+export type ActionsPanelTestHarness = ReturnType<typeof createActionsPanelGame>;
